+++ crates/l2/starknet-eth-bridge-client/src/clients/eth_bridge.rs
@@ -4,12 +4,12 @@ use crate::interfaces::eth_bridge::StarknetEthBridge;
 use utils::{LocalWalletSignerMiddleware, StarknetContractClient};
 
 use ethers::types::Address;
-use starknet_proxy_client::clients::proxy_3_0_2::ProxySupport;
+use starknet_proxy_client::clients::proxy_3_0_2::ProxySupport3_0_2;
 
 /// Client to interact with a Starknet Eth Bridge
 pub struct StarknetEthBridgeContractClient {
     eth_bridge: StarknetEthBridge<LocalWalletSignerMiddleware>,
-    proxy_support: ProxySupport<LocalWalletSignerMiddleware>,
+    proxy_support: ProxySupport3_0_2<LocalWalletSignerMiddleware>,
     eth_bridge_implementation: Address,
 }
 
@@ -21,7 +21,7 @@ impl StarknetEthBridgeContractClient {
     ) -> Self {
         Self {
             eth_bridge: StarknetEthBridge::new(address, client.clone()),
-            proxy_support: ProxySupport::new(address, client.clone()),
+            proxy_support: ProxySupport3_0_2::new(address, client.clone()),
             eth_bridge_implementation: implementation_address,
         }
     }
@@ -33,8 +33,8 @@ impl AsRef<StarknetEthBridge<LocalWalletSignerMiddleware>> for StarknetEthBridge
     }
 }
 
-impl AsRef<ProxySupport<LocalWalletSignerMiddleware>> for StarknetEthBridgeContractClient {
-    fn as_ref(&self) -> &ProxySupport<LocalWalletSignerMiddleware> {
+impl AsRef<ProxySupport3_0_2<LocalWalletSignerMiddleware>> for StarknetEthBridgeContractClient {
+    fn as_ref(&self) -> &ProxySupport3_0_2<LocalWalletSignerMiddleware> {
         &self.proxy_support
     }
 }
