+++ crates/l2/starknet-proxy-client/src/clients/mod.rs
@@ -1,3 +1,2 @@
-pub mod proxy;
 pub mod proxy_3_0_2;
 pub mod proxy_5_0_0;
