+++ crates/l2/starknet-proxy-client/src/interfaces/mod.rs
@@ -1,0 +1,1 @@
+pub mod proxy;
