+++ crates/l2/starknet-proxy-client/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod clients;
 pub mod deploy;
+pub mod interfaces;
