+++ crates/l2/starknet-token-bridge-client/src/clients/token_bridge.rs
@@ -4,12 +4,12 @@ use crate::interfaces::token_bridge::StarknetTokenBridge;
 use utils::{LocalWalletSignerMiddleware, StarknetContractClient};
 
 use ethers::types::Address;
-use starknet_proxy_client::clients::proxy_5_0_0::ProxySupportLatest;
+use starknet_proxy_client::clients::proxy_5_0_0::ProxySupport5_0_0;
 
 /// Client to interact with a Token Bridge (ERC20)
 pub struct StarknetTokenBridgeContractClient {
     token_bridge: StarknetTokenBridge<LocalWalletSignerMiddleware>,
-    proxy_support: ProxySupportLatest<LocalWalletSignerMiddleware>,
+    proxy_support: ProxySupport5_0_0<LocalWalletSignerMiddleware>,
     token_bridge_implementation: Address,
 }
 
@@ -21,7 +21,7 @@ impl StarknetTokenBridgeContractClient {
     ) -> Self {
         Self {
             token_bridge: StarknetTokenBridge::new(address, client.clone()),
-            proxy_support: ProxySupportLatest::new(address, client.clone()),
+            proxy_support: ProxySupport5_0_0::new(address, client.clone()),
             token_bridge_implementation: implementation_address,
         }
     }
@@ -33,8 +33,8 @@ impl AsRef<StarknetTokenBridge<LocalWalletSignerMiddleware>> for StarknetTokenBr
     }
 }
 
-impl AsRef<ProxySupportLatest<LocalWalletSignerMiddleware>> for StarknetTokenBridgeContractClient {
-    fn as_ref(&self) -> &ProxySupportLatest<LocalWalletSignerMiddleware> {
+impl AsRef<ProxySupport5_0_0<LocalWalletSignerMiddleware>> for StarknetTokenBridgeContractClient {
+    fn as_ref(&self) -> &ProxySupport5_0_0<LocalWalletSignerMiddleware> {
         &self.proxy_support
     }
 }
